Fully functioning finances_cancel_last command.

+++ src/main/java/com/space_asians/w1ncent/entities/Member.java
@@ -16,5 +16,5 @@ public class Member {
     private String name;
 
     @Column
-    int balance;
+    float balance;
 }
